Moved: Implementation from Core to Factory
